requirements.txt
astroid==1.6.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.11.5
future==0.16.0
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.8.1
python-telegram-bot==9.0.0
six==1.11.0
wrapt==1.10.11
youtube-dl==2017.12.31
